Refactor votes received implementation
