--- services/blockchain-indexer/shared/dataService/voters.js
@@ -1,41 +1,0 @@
-/*
- * LiskHQ/lisk-service
- * Copyright © 2022 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- *
- */
-const dataService = require('./business');
-
-const getVoters = async params => {
-	const voters = {
-		data: [],
-		meta: {},
-	};
-
-	const response = await dataService.getVoters(params);
-	voters.data = response.data.voters ? response.data.voters : response.data;
-	voters.meta = {
-		limit: response.meta.limit,
-		count: response.data.voters ? response.data.voters.length : response.meta.count,
-		offset: response.meta.offset,
-		total: response.data.voteCount || response.meta.total,
-		address: response.data.address,
-		publicKey: response.data.publicKey,
-		username: response.data.username,
-	};
-
-	return voters;
-};
-
-module.exports = {
-	getVoters,
-};
